fix compiler warning
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@80449 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/igblast/igblastn_app.cpp
@@ -581,7 +581,7 @@ int CIgBlastnApp::Run(void)
 	    string delim = ", ";
 	    NStr::Split( sra_run_accessions, delim, run_arr );
 
-	    for( int ndx=0; ndx<run_arr.size(); ndx ++){
+	    for( int ndx=0; ndx < (int)run_arr.size(); ndx ++){
 		string one_run = NStr::TruncateSpaces( run_arr[ndx]);
 		all_sra_runs.push_back( one_run );
 	    }
